+++ NEWS.md
@@ -4,6 +4,7 @@
 * `VCF2RADdata` now can make use of .csi indices on VCFs that are already bgzipped.
 * Bug fix in `Export_MAPpoly` relating to how ploidy is coded.
 * Bug fix in `AddAlleleFreqByTaxa` if only one PC is used.
+* Update `readStacks` for new file formats.
 
 # polyRAD 2.0